fixes issue #26
* changed twitter user_id to BIGINT
* fixed bug in creetdao

+++ src/main/resources/db/migration/V2__Twitter_Integration.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `critter`.`GroupTwitterCredentials` (
   `group_id` VARCHAR(36) NOT NULL,
   `screen_name` VARCHAR(15) NOT NULL,
-  `user_id` INT NOT NULL,
+  `user_id` BIGINT NOT NULL,
   `token` VARCHAR(64) NOT NULL,
   `token_secret` VARCHAR(64) NOT NULL,
   PRIMARY KEY (`group_id`)
